Scale new fitness by population size
This fixes 2 bugs that ended up being related to each other: population shrinking sometimes and index error on crossover sometimes.  The former was due to the chance that the shuffled fps sometimes didn't have as many entries as there were individuals in the population.  I have scaled the fitnesses such that the shuffled list is at least guarenteed to be large enough.  The second issue I came to realize happened when the population was shrunk to an odd number (crossover assumes even number of parents, as it goes by 2).  Preventing the shrinking fixed this problem.